FSTR/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1
defusedxml==0.7.1
django-allauth==0.52.0
django-filter==22.1
django-rest-framework==0.1.0
Django==4.1.7
djangorestframework==3.14.0
idna==3.4
Markdown==3.4.1
oauthlib==3.2.2
pip==23.0
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2022.7.1
requests-oauthlib==1.3.1
requests==2.28.2
setuptools==60.2.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
wheel==0.37.1
drf-yasg
django-rest-swagger